+++ RCS.DIS.DataServices/DataModel/Zorgactiviteit.cs
@@ -22,7 +22,7 @@ namespace RCS.DIS.DataServices.DataModel
     
         public string ZorgactiviteitCode { get; set; }
         public string Omschrijving { get; set; }
-        public short ZorgprofielklasseCode { get; set; }
+        public int ZorgprofielklasseCode { get; set; }
         public System.DateTime Peildatum { get; set; }
         public Nullable<System.DateTime> Bestandsdatum { get; set; }
         public string Versie { get; set; }
